Remove obsolete getContentAnchorEl prop from LookupFilter MenuProps (#676)
In MUI v5, the `getContentAnchorEl` prop was removed
https://mui.com/material-ui/migration/v5-component-changes/#remove-getcontentanchorel-prop

So with this prop still being supplied via MenuProps currently, it leads
to the following console error whenever you click on a filter menu from
@material-table/core:

> Warning: React does not recognize the `getContentAnchorEl` prop on a
DOM element. If you intentionally want it to appear in the DOM as a
custom attribute, spell it as lowercase `getcontentanchorel` instead. If
you accidentally passed it from a parent component, remove it from the
DOM element.

The prop isn't doing anything anymore, and the filter menu still works
without it, so it seems it should just be removed in the MUI v5 variant
of @material-table/core.

Co-authored-by: Steven DeMartini <sjdemartini@users.noreply.github.com>