Fix values used for aria-sort in MTableHeader.
The previous code used 'Ascendant' and 'Descendant' as the aria-sort
value.  These are not in spec (https://w3c.github.io/aria/#aria-sort).
The correct values are 'ascending' and 'descending'.

Also fix the prop-types declarations for columns.ariaSortAsc and
columns.ariaSortDesc to match the spec (and the declaration in
React.AriaAttributes).