fix: add string array to columns.field prop types
Support for passing an array to the 'field' attribute of a column was
previously added to allow the selection of keys with full-stops.
Although the the type of the field was updated in the type declaration
file, 'prop-types.js' was not updated to reflect these changes,
resulting in runtime errors.